refactor map container
