+++ lib/widgets/your_location_button.dart
@@ -1,15 +1,42 @@
 import 'package:flutter/material.dart';
+import 'package:geolocator/geolocator.dart';
+import 'package:trufi_core/blocs/gps_location/location_provider_cubit.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:trufi_core/widgets/alerts.dart';
+import 'package:trufi_core/widgets/map/trufi_map_controller.dart';
 
-class YourLocationButton extends StatelessWidget {
-  const YourLocationButton({this.onPressed, Key key}): super(key: key);
+import 'map/trufi_map_controller.dart';
 
-  final VoidCallback onPressed;
+class YourLocationButton extends StatefulWidget {
+  const YourLocationButton({@required this.trufiMapController});
+  final TrufiMapController trufiMapController;
 
+  @override
+  _YourLocationButtonState createState() => _YourLocationButtonState();
+}
+
+class _YourLocationButtonState extends State<YourLocationButton>
+    with TickerProviderStateMixin {
   @override
   Widget build(BuildContext context) {
     return FloatingActionButton(
       backgroundColor: Theme.of(context).backgroundColor,
-      onPressed: onPressed,
+      onPressed: () async {
+        try {
+          final location =
+              context.read<LocationProviderCubit>().state.currentLocation;
+          widget.trufiMapController.moveToYourLocation(
+            location: location,
+            context: context,
+            tickerProvider: this,
+          );
+        } on PermissionDeniedException catch (_) {
+          showDialog(
+            context: context,
+            builder: (context) => buildAlertLocationServicesDenied(context),
+          );
+        }
+      },
       heroTag: null,
       child: const Icon(
         Icons.my_location,
